# [SQL Storage Indexer] Skip unmarshalling JSON fields in statics, artifacts and static handlers