--- a/artifacts.go
+++ b/artifacts.go
@@ -82,6 +82,8 @@ func (h *artifactsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 
 	opts := packages.NameVersionFilter(packageName, packageVersion)
+	opts.SkipJSON = true
+
 	pkgs, err := h.indexer.Get(r.Context(), &opts)
 	if err != nil {
 		logger.Error("getting package path failed",