--- a/internal/database/repository.go
+++ b/internal/database/repository.go
@@ -22,4 +22,5 @@ type Repository interface {
 type WhereOptions interface {
 	Where() (string, []any)
 	UseFullData() bool
+	SkipJSONFields() bool
 }