--- a/packages/package.go
+++ b/packages/package.go
@@ -247,6 +247,10 @@ func MustParsePackage(basePath string, fsBuilder FileSystemBuilder) (*Package, e
 	return p, nil
 }
 
+// NewMinimalPackage creates a new minimal package instance with the given name, version, and spec version
+// without loading any files from disk. It also sets runtime fields like versionSemVer and specMajorMinorSemVer.
+// It is used in scenarios where only basic package information is needed.
+// The returned package instance will not have fields like Assets, DataStreams, PolicyTemplates, etc. populated.
 func NewMinimalPackage(name, packageVersion, specVersion string) (*Package, error) {
 	p := Package{
 		FormatVersion: specVersion,
@@ -260,60 +264,6 @@ func NewMinimalPackage(name, packageVersion, specVersion string) (*Package, erro
 	}
 	p.Release = releaseForSemVerCompat(p.versionSemVer)
 	return &p, nil
-}
-
-type PackageOption func(p *Package) error
-
-func NewPackageWithOpts(logger *zap.Logger, opts ...PackageOption) (*Package, error) {
-	p := Package{}
-	for _, opt := range opts {
-		if err := opt(&p); err != nil {
-			return nil, err
-		}
-	}
-	return &p, nil
-
-}
-
-func WithPackageName(packageName string) PackageOption {
-	return func(p *Package) error {
-		p.Name = packageName
-		return nil
-	}
-}
-
-func WithPackageVersion(packageVersion string) PackageOption {
-	return func(p *Package) error {
-		p.Version = packageVersion
-		var err error
-		p.versionSemVer, err = semver.StrictNewVersion(packageVersion)
-		if err != nil {
-			return err
-		}
-		if p.Release == "" {
-			p.Release = releaseForSemVerCompat(p.versionSemVer)
-		}
-		return nil
-	}
-}
-
-func WithFormatVersion(specVersion string) PackageOption {
-	return func(p *Package) error {
-		p.FormatVersion = specVersion
-
-		specSemVer, err := semver.StrictNewVersion(p.FormatVersion)
-		if err != nil {
-			return fmt.Errorf("invalid format spec version '%s': %w", p.FormatVersion, err)
-		}
-
-		specMajorMinorVersion := fmt.Sprintf("%d.%d.0", specSemVer.Major(), specSemVer.Minor())
-
-		p.specMajorMinorSemVer, err = semver.StrictNewVersion(specMajorMinorVersion)
-		if err != nil {
-			return fmt.Errorf("invalid format spec version '%s': %w", specMajorMinorVersion, err)
-		}
-		return nil
-	}
 }
 
 // NewPackage creates a new package instances based on the given base path.