--- a/static.go
+++ b/static.go
@@ -73,7 +73,7 @@ func (h *staticHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 
 	opts := packages.NameVersionFilter(params.packageName, params.packageVersion)
-	opts.SkipJSON = true
+	opts.SkipPackageData = true
 
 	pkgs, err := h.indexer.Get(r.Context(), &opts)
 	if err != nil {