Closes https://github.com/elastic/package-registry/issues/1387

This PR allows to not read the JSON field from the database for requests received to statics, artifacts or signature handlers (`/package` and `/epr` endpoints):
- `/package/<package>/<version>/<static_file_path>` (staticsHandler)
- `/epr/<package>/<package>-<version>.zip` (artifactsHandler)
- `/epr/<package>/<package>-<version>.zip.sig` (signaturesHandler)

The required responses could be built without the information in `data` or `BaseData` columns from the SQL database (JSON blobs).

Results of a local benchmark for the specific Get calls executed in the API endpoints mentioned above:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/package-registry/internal/storage
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                                    │ sql_prod_get_statics_main.txt │   sql_prod_statics_skip_json.txt    │
                                    │            sec/op             │   sec/op     vs base                │
SQLIndexerGetStaticsAndArtifacts-16                     998.9µ ± 7%   247.1µ ± 4%  -75.27% (p=0.000 n=10)

                                    │ sql_prod_get_statics_main.txt │    sql_prod_statics_skip_json.txt    │
                                    │             B/op              │     B/op      vs base                │
SQLIndexerGetStaticsAndArtifacts-16                   171.81Ki ± 0%   28.22Ki ± 0%  -83.58% (p=0.000 n=10)

                                    │ sql_prod_get_statics_main.txt │   sql_prod_statics_skip_json.txt   │
                                    │           allocs/op           │ allocs/op   vs base                │
SQLIndexerGetStaticsAndArtifacts-16                     1612.0 ± 0%   404.0 ± 0%  -74.94% (p=0.000 n=10)
```